+++ collective/notifications/external.py
@@ -1,5 +1,5 @@
 import email
-
+from Acquisition import aq_base
 from plone import api
 from zope.interface import implements
 
@@ -12,8 +12,14 @@ class EmailNotifier(object):
 
     def send(self, notification):
         portal = api.portal.get()
-        subject = "Notification from {}".format(portal.title)
-        msg = email.message_from_string(notification.note)
+        base_notification = aq_base(notification)
+        subject = getattr(base_notification, 'email_subject', None)
+        if not subject:
+            subject = "Notification from {}".format(portal.title)
+        email_body = getattr(base_notification, 'email_body', None)
+        if not email_body:
+            email_body = notification.note
+        msg = email.message_from_string(email_body)
         msg.set_charset('utf-8')
         name = api.portal.get_registry_record('plone.email_from_name')
         address = api.portal.get_registry_record('plone.email_from_address')
@@ -22,7 +28,7 @@ class EmailNotifier(object):
         for recipient in notification.recipients:
             user = api.user.get(userid=recipient)
             address = user.getProperty('email', None)
-            if address is None:
+            if not address:
                 continue
             mailhost.send(msg,
                           subject=subject,