+++ collective/notifications/notifications.py
@@ -121,7 +121,9 @@ class Notification(Persistent):
                  user=None,
                  url=None,
                  first_read=False,
-                 external=None):
+                 external=None,
+                 email_body=None,
+                 email_subject=None):
         uid = getUtility(IUUIDGenerator)()
         context_uid = getattr(context, 'UID', False) and context.UID() or context.id
         self.uid = uid
@@ -131,6 +133,8 @@ class Notification(Persistent):
         self.recipients = self.get_recipients(recipients)
         self.first_read = first_read
         self.external = external
+        self.email_body = email_body
+        self.email_subject = email_subject
         if user is None:
             user = api.user.get_current()
             if user is not None:
@@ -190,7 +194,9 @@ def handle_notification_requested(event):
                                 event.user,
                                 event.url,
                                 event.first_read,
-                                event.external)
+                                event.external,
+                                email_body=event.email_body,
+                                email_subject=event.email_subject)
     site = getSite()
     storage = INotificationStorage(site)
     storage.add_notification(notification)